chore: move conversion logic to new file
